Pause menu now displays
Added our first version of the pause menu. Fixed the bug were the pause menu was not displaying. The buttons should be functional, however I do not know how to disable the camera and turn on a cursor so we cannot know if the buttons work until we do that.