Added a test scene for invenotry UI
In this test scene I added a simple invendtory UI. I did this by making a copy of Andrew's Item Pickup Test Scene. In it I have a custom HUD for the inventory with 9 slots, 2 of them were built in to the MasterInvenotry asset, 1 slot is occupied by a custom panel and the other 6 are empty.