Repurposed the scene management
Now the Exit button in the pause menu goes back to the main menu. I didn't delete the function that simply closes the application, it's still there if we want to use that one instead. I need the new function to test the save and load functions. I'll explain in more detail later on.