Fixed a bug
I just realized that it's impossible to free the cursor once time is resumed. So, I'm gonna have to sacrifice having a visible cursor once the exit button is clicked, because we need time to be resumed to load scenes, at least for now. Once you hit the escape key in the main meny, the cursor is free but you can't see it. However, the buttons will still hiighlight when hovered over, so that's how you'll know where you're at.